FogBugz #110927: Fixing Yocto build warning messages - 4 uboot and 1 gator
For gator this was a matter of figuring out the INSANE skip syntax
For uboot there were 2 changes made. The first was to indicate that we do not need to build a uboot package because uboot isn't something that gets installed. The second patch put files in the correct location