FogBugz #190908: Update the kernel recipe for AMP branch to pickup kernel from AMP branch
This is to update the recipes for kernel AMP branch so that it can will use
kernel from AMP branch. Also update the build so that it can dynamically
select the GIT tree from environment.
Also remove the patching for kernel since we are now releasing the AMP
reference design from GIT, not as a patch.

Signed-off-by: Tien Hock Loh <thloh@altera.com>